Make repositories have meaningful names when publishing
Also:
 * Remove deprecations.
 * Encapsulate `snapshots` and `releases` properties.
 * Introduce `repo` package for corresponding types.
 * Pull up `Repository`, `Credentials`, and `RepoSlug` to the package level.
