+++ buildSrc/src/main/kotlin/BuildExtensions.kt
@@ -36,7 +36,7 @@ import io.spine.dependency.local.ProtoData
 import io.spine.dependency.local.ProtoTap
 import io.spine.dependency.test.Kotest
 import io.spine.dependency.test.Kover
-import io.spine.gradle.standardToSpineSdk
+import io.spine.gradle.repo.standardToSpineSdk
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.tasks.JavaExec
@@ -67,7 +67,7 @@ import org.gradle.plugin.use.PluginDependencySpec
 private const val ABOUT_DEPENDENCY_EXTENSIONS = ""
 
 /**
- * Applies [standard][standardToSpineSdk] repositories to this `buildscript`.
+ * Applies [standard][io.spine.gradle.repo.standardToSpineSdk] repositories to this `buildscript`.
  */
 fun ScriptHandlerScope.standardSpineSdkRepositories() {
     repositories.standardToSpineSdk()
