+++ buildSrc/src/main/kotlin/io/spine/gradle/publish/CloudRepo.kt
@@ -26,7 +26,7 @@
 
 package io.spine.gradle.publish
 
-import io.spine.gradle.Repository
+import io.spine.gradle.repo.Repository
 
 /**
  * CloudRepo Maven repository.
