+++ buildSrc/src/main/kotlin/io/spine/gradle/publish/CustomPublicationHandler.kt
@@ -26,7 +26,7 @@
 
 package io.spine.gradle.publish
 
-import io.spine.gradle.Repository
+import io.spine.gradle.repo.Repository
 import org.gradle.api.Project
 import org.gradle.api.publish.maven.MavenPublication
