+++ buildSrc/src/main/kotlin/io/spine/gradle/repo/Credentials.kt
@@ -24,43 +24,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.gradle
-
-import org.gradle.api.GradleException
+package io.spine.gradle.repo
 
 /**
- * A name of a repository.
+ * Password credentials for a Maven repository.
  */
-class RepoSlug(val value: String) {
-
-    companion object {
-
-        /**
-         * The name of the environment variable containing the repository slug, for which
-         * the Gradle build is performed.
-         */
-        private const val environmentVariable = "REPO_SLUG"
-
-        /**
-         * Reads `REPO_SLUG` environment variable and returns its value.
-         *
-         * In case it is not set, a [GradleException] is thrown.
-         */
-        fun fromVar(): RepoSlug {
-            val envValue = System.getenv(environmentVariable)
-            if (envValue.isNullOrEmpty()) {
-                throw GradleException("`REPO_SLUG` environment variable is not set.")
-            }
-            return RepoSlug(envValue)
-        }
-    }
-
-    override fun toString(): String = value
-
-    /**
-     * Returns the GitHub URL to the project repository.
-     */
-    fun gitHost(): String {
-        return "git@github.com-publish:${value}.git"
-    }
-}
+data class Credentials(
+    val username: String?,
+    val password: String?
+)
